+++ packages/examples/src/langium/reactStatemachine.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { MonacoEditorReactComp } from '@typefox/monaco-editor-react';
-import { createLangiumGlobalConfig } from './langiumWrapperConfig.js';
+import { createLangiumGlobalConfig } from './config/wrapperStatemachineConfig.js';
 
 import { buildWorkerDefinition } from 'monaco-editor-workers';
 buildWorkerDefinition('../../../../node_modules/monaco-editor-workers/dist/workers', import.meta.url, false);